# Fix concurrent MemoryScope.close() or MemoryScope.dup() race