In case a MemorySegment is not confined to an "owner" thread, it is possible to call close() or withOwnerThread(newThread) on it from multiple threads concurrently and such calls are not screened for thread confinement but just forwarded to MemoryScope.close() and .dup(). It is therefore vital that close() and dup() are made atomic so that cleanupAction is executed just once.
<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must not contain extraneous whitespace
- [ ] Change must be properly reviewed

### Download
`$ git fetch https://git.openjdk.java.net/panama-foreign pull/160/head:pull/160`
`$ git checkout pull/160`