@plevart 2020-05-11
The ParallelSum benchmark doesn't show and regressions:
```
before fix:

Benchmark                                         Mode  Cnt     Score     Error  Units
ParallelSum.segment_parallel                      avgt   30    42.298 ±   0.139  ms/op
ParallelSum.segment_parallel_bulk                 avgt   30    38.268 ±   0.318  ms/op
ParallelSum.segment_serial                        avgt   30   109.087 ±   0.618  ms/op
ParallelSum.segment_stream_findany_parallel       avgt   30  1721.486 ± 306.951  ms/op
ParallelSum.segment_stream_findany_parallel_bulk  avgt   30    32.263 ±   1.572  ms/op
ParallelSum.segment_stream_findany_serial         avgt   30  5090.839 ±  19.335  ms/op
ParallelSum.segment_stream_findany_serial_bulk    avgt   30   105.765 ±   1.527  ms/op
ParallelSum.segment_stream_parallel               avgt   30    38.995 ±   0.236  ms/op
ParallelSum.segment_stream_parallel_bulk          avgt   30    38.146 ±   0.225  ms/op
ParallelSum.unsafe_parallel                       avgt   30     6.476 ±   0.226  ms/op
ParallelSum.unsafe_serial                         avgt   30   107.892 ±   0.882  ms/op

after fix:

Benchmark                                         Mode  Cnt     Score     Error  Units
ParallelSum.segment_parallel                      avgt   30    42.195 ±   0.094  ms/op
ParallelSum.segment_parallel_bulk                 avgt   30    38.216 ±   0.269  ms/op
ParallelSum.segment_serial                        avgt   30    98.562 ±   9.357  ms/op
ParallelSum.segment_stream_findany_parallel       avgt   30  1532.052 ± 325.223  ms/op
ParallelSum.segment_stream_findany_parallel_bulk  avgt   30    32.167 ±   1.664  ms/op
ParallelSum.segment_stream_findany_serial         avgt   30  5443.200 ± 370.581  ms/op
ParallelSum.segment_stream_findany_serial_bulk    avgt   30   104.543 ±   0.924  ms/op
ParallelSum.segment_stream_parallel               avgt   30    39.010 ±   0.281  ms/op
ParallelSum.segment_stream_parallel_bulk          avgt   30    38.161 ±   0.314  ms/op
ParallelSum.unsafe_parallel                       avgt   30     6.386 ±   0.250  ms/op
ParallelSum.unsafe_serial                         avgt   30    99.642 ±   8.711  ms/op
```