@mcimadamore 2020-05-13
> I think this should replace current version. I'm all for it. Good work.

Cool! Then I suggest you maybe close this PR and I will submit another with the new impl, if that's ok?