@plevart 2020-05-13
Closing this PR in favor of Maurizio's implementation.